requirements.txt
# Dependencias del proyecto La Casa de los Abuelos

# Framework IA Multiagente
crewai==0.45.0

# LLM APIs
ollama==0.3.0
google-generativeai==0.8.3

# Manejo de environment variables
python-dotenv==1.0.0

# Database
sqlite3  # Viene con Python, pero aquí va para claridad

# Logging y monitoreo
python-json-logger==2.1.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Formateo y linting
black==24.3.0
flake8==7.0.0
pylint==3.0.0

# Utilidades
requests==2.31.0
pydantic==2.6.0

# Opcional: para CLI mejorada
click==8.1.7
rich==13.7.0

# Opcional: para análisis de datos
pandas==2.1.0
